+++ Rdmp.UI/SimpleDialogs/UserSettingsUI.cs
@@ -84,6 +84,7 @@ namespace Rdmp.UI.SimpleDialogs
             cbScoreZeroForCohortAggregateContainers.Checked = UserSettings.ScoreZeroForCohortAggregateContainers;
             cbAdvancedFindFilters.Checked = UserSettings.AdvancedFindFilters;
             cbIncludeZeroSeriesInGraphs.Checked = UserSettings.IncludeZeroSeriesInGraphs;
+            cbAlwaysJoinEverything.Checked = UserSettings.AlwaysJoinEverything;
             tbCreateDatabaseTimeout.Text = UserSettings.CreateDatabaseTimeout.ToString();
             tbArchiveTriggerTimeout.Text = UserSettings.ArchiveTriggerTimeout.ToString();
             tbTooltipAppearDelay.Text = UserSettings.TooltipAppearDelay.ToString();
@@ -102,6 +103,7 @@ namespace Rdmp.UI.SimpleDialogs
             AddTooltip(cbScoreZeroForCohortAggregateContainers,nameof(UserSettings.ScoreZeroForCohortAggregateContainers));
             AddTooltip(cbAdvancedFindFilters,nameof(UserSettings.AdvancedFindFilters));
             AddTooltip(cbIncludeZeroSeriesInGraphs,nameof(UserSettings.IncludeZeroSeriesInGraphs));
+            AddTooltip(cbAlwaysJoinEverything,nameof(UserSettings.AlwaysJoinEverything));
             AddTooltip(label7, nameof(UserSettings.CreateDatabaseTimeout));
             AddTooltip(tbCreateDatabaseTimeout, nameof(UserSettings.CreateDatabaseTimeout));
             AddTooltip(label13, nameof(UserSettings.ArchiveTriggerTimeout));
@@ -212,6 +214,9 @@ namespace Rdmp.UI.SimpleDialogs
 
             if (cb == cbIncludeZeroSeriesInGraphs)
                 UserSettings.IncludeZeroSeriesInGraphs = cb.Checked;
+
+            if (cb == cbAlwaysJoinEverything)
+                UserSettings.AlwaysJoinEverything = cb.Checked;
         }
 
         private void ddTheme_SelectedIndexChanged(object sender, EventArgs e)